refactor: update the app layer to use the new auth module
Update the app module so that it makes use of the new Server:Auth module
