+++ app/build.gradle
@@ -52,6 +52,7 @@ android {
         exclude "META-INF/AL2.0"
         exclude "META-INF/LGPL2.1"
         exclude "META-INF/library-core_release.kotlin_module"
+        exclude "META-INF/auth_debug.kotlin_module"
         exclude "META-INF/search_debug.kotlin_module"
         exclude "META-INF/trending_debug.kotlin_module"
         exclude "META-INF/user_debug.kotlin_module"
@@ -80,11 +81,13 @@ dependencies {
     implementation project(path: ":libraries:database")
     implementation project(path: ":libraries:encryption")
     implementation project(path: ":libraries:kitsu")
+    implementation project(path: ":libraries:kitsu:auth")
     implementation project(path: ":libraries:kitsu:search")
     implementation project(path: ":libraries:kitsu:trending")
     implementation project(path: ":libraries:kitsu:user")
     implementation project(path: ":libraries:library")
     implementation project(path: ":libraries:server")
+    implementation project(path: ":libraries:server:auth")
     implementation project(path: ":libraries:server:search")
     implementation project(path: ":libraries:server:trending")
     implementation project(path: ":libraries:server:user")
