+++ app/src/main/java/com/chesire/nekome/injection/KitsuModule.kt
@@ -1,8 +1,6 @@
 package com.chesire.nekome.injection
 
-import com.chesire.nekome.kitsu.api.auth.KitsuAuth
 import com.chesire.nekome.kitsu.api.library.KitsuLibrary
-import com.chesire.nekome.server.api.AuthApi
 import com.chesire.nekome.server.api.LibraryApi
 import dagger.Binds
 import dagger.Module
@@ -15,12 +13,6 @@ import dagger.hilt.android.components.ApplicationComponent
 @Module
 @InstallIn(ApplicationComponent::class)
 abstract class KitsuModule {
-    /**
-     * Binds the instance of [KitsuAuth] to [AuthApi].
-     */
-    @Binds
-    abstract fun providesAuthApi(kitsuAuth: KitsuAuth): AuthApi
-
     /**
      * Binds the instance of [KitsuLibrary] to [LibraryApi].
      */
