+++ app/src/main/java/com/chesire/nekome/services/RefreshAuthWorker.kt
@@ -6,8 +6,8 @@ import androidx.hilt.work.WorkerInject
 import androidx.work.CoroutineWorker
 import androidx.work.WorkerParameters
 import com.chesire.nekome.account.UserRepository
+import com.chesire.nekome.auth.api.AuthApi
 import com.chesire.nekome.core.Resource
-import com.chesire.nekome.server.api.AuthApi
 import timber.log.Timber
 
 /**
