+++ features/login/src/main/java/com/chesire/nekome/app/login/details/DetailsViewModel.kt
@@ -5,8 +5,8 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.chesire.nekome.account.UserRepository
+import com.chesire.nekome.auth.api.AuthApi
 import com.chesire.nekome.core.Resource
-import com.chesire.nekome.server.api.AuthApi
 import com.hadilq.liveevent.LiveEvent
 import kotlinx.coroutines.launch
 import timber.log.Timber
